@@ -16,7 +16,7 @@
         "source": "https://github.com/hyperf/hyperf"
     },
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.0",
         "hyperf/contract": "~3.0.0",
         "hyperf/utils": "~3.0.0",
         "psr/container": "^1.0|^2.0",
